requirements.txt
crewai
langchain_google_genai
load_dotenv
streamlit crewai
crewai_tools